Move RegexCache to internal subdirectory
