+++ java/libphonenumber/src/com/google/i18n/phonenumbers/AsYouTypeFormatter.java
@@ -18,6 +18,7 @@ package com.google.i18n.phonenumbers;
 
 import com.google.i18n.phonenumbers.Phonemetadata.NumberFormat;
 import com.google.i18n.phonenumbers.Phonemetadata.PhoneMetadata;
+import com.google.i18n.phonenumbers.internal.RegexCache;
 
 import java.util.ArrayList;
 import java.util.Iterator;
