+++ java/libphonenumber/src/com/google/i18n/phonenumbers/PhoneNumberUtil.java
@@ -21,6 +21,7 @@ import com.google.i18n.phonenumbers.Phonemetadata.PhoneMetadata;
 import com.google.i18n.phonenumbers.Phonemetadata.PhoneNumberDesc;
 import com.google.i18n.phonenumbers.Phonenumber.PhoneNumber;
 import com.google.i18n.phonenumbers.Phonenumber.PhoneNumber.CountryCodeSource;
+import com.google.i18n.phonenumbers.internal.RegexCache;
 
 import java.util.ArrayList;
 import java.util.Arrays;
