+++ java/libphonenumber/src/com/google/i18n/phonenumbers/internal/RegexBasedMatcher.java
@@ -17,7 +17,6 @@
 package com.google.i18n.phonenumbers.internal;
 
 import com.google.i18n.phonenumbers.Phonemetadata.PhoneNumberDesc;
-import com.google.i18n.phonenumbers.RegexCache;
 
 import java.util.regex.Matcher;
