+++ java/libphonenumber/src/com/google/i18n/phonenumbers/internal/RegexCache.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.i18n.phonenumbers;
+package com.google.i18n.phonenumbers.internal;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
