+++ java/libphonenumber/test/com/google/i18n/phonenumbers/internal/RegexCacheTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.i18n.phonenumbers;
+package com.google.i18n.phonenumbers.internal;
 
 import junit.framework.TestCase;
